2023.04.18 이예은 코딩

+++ src/main/java/Code_Up/Q_1091.java
@@ -6,7 +6,7 @@ public class Q_1091 {
     public static void main(String[] args) {
         Scanner sc = new Scanner(System.in);
 
-        int a = sc.nextInt();
+        long a = sc.nextLong();
         int m = sc.nextInt();
         int d = sc.nextInt();
         int n = sc.nextInt();
@@ -14,6 +14,6 @@ public class Q_1091 {
         for(int i = 1; i<n; i++){
             a= a*m +d;
         }
-        System.out.print((long)a);
+        System.out.print(a);
     }
 }
